src/collaboration.py
```python
# Importing necessary libraries
from git import Repo
from github import Github
from gitlab import Gitlab
from bitbucket.bitbucket import Bitbucket

class Collaboration:
    def __init__(self, repo_url, platform, token):
        self.repo_url = repo_url
        self.platform = platform
        self.token = token

    def git_version_control(self):
        """
        Function to clone a repository using Git
        """
        repo = Repo.clone_from(self.repo_url, './repo')
        return repo

    def github_collaboration(self):
        """
        Function to collaborate using GitHub
        """
        g = Github(self.token)
        user = g.get_user()
        repo = user.get_repo(self.repo_url)
        return repo

    def gitlab_ci_cd(self):
        """
        Function to use CI/CD pipelines in GitLab
        """
        gl = Gitlab(self.repo_url, private_token=self.token)
        project = gl.projects.get(self.repo_url)
        pipelines = project.pipelines.list()
        return pipelines

    def bitbucket_repository_management(self):
        """
        Function to manage repositories in Bitbucket
        """
        bb = Bitbucket(self.token)
        repo = bb.repository(self.repo_url)
        return repo
```
