tests/test_collaboration.py
```python
# Importing necessary libraries
import unittest
from unittest.mock import patch, MagicMock
from src.collaboration import Collaboration

class TestCollaboration(unittest.TestCase):
    @patch('src.collaboration.Repo')
    def test_git_version_control(self, mock_repo):
        """
        Test to check the git_version_control function
        """
        collaboration = Collaboration('https://github.com/user/repo', 'github', 'token')
        collaboration.git_version_control()
        mock_repo.clone_from.assert_called_once_with('https://github.com/user/repo', './repo')

    @patch('src.collaboration.Github')
    def test_github_collaboration(self, mock_github):
        """
        Test to check the github_collaboration function
        """
        mock_github.return_value.get_user.return_value.get_repo.return_value = MagicMock()
        collaboration = Collaboration('https://github.com/user/repo', 'github', 'token')
        collaboration.github_collaboration()
        mock_github.assert_called_once_with('token')

    @patch('src.collaboration.Gitlab')
    def test_gitlab_ci_cd(self, mock_gitlab):
        """
        Test to check the gitlab_ci_cd function
        """
        mock_gitlab.return_value.projects.get.return_value.pipelines.list.return_value = MagicMock()
        collaboration = Collaboration('https://gitlab.com/user/repo', 'gitlab', 'token')
        collaboration.gitlab_ci_cd()
        mock_gitlab.assert_called_once_with('https://gitlab.com/user/repo', private_token='token')

    @patch('src.collaboration.Bitbucket')
    def test_bitbucket_repository_management(self, mock_bitbucket):
        """
        Test to check the bitbucket_repository_management function
        """
        mock_bitbucket.return_value.repository.return_value = MagicMock()
        collaboration = Collaboration('https://bitbucket.org/user/repo', 'bitbucket', 'token')
        collaboration.bitbucket_repository_management()
        mock_bitbucket.assert_called_once_with('token')

if __name__ == '__main__':
    unittest.main()
```
